"ML System Architecture for Operation Stability"
The author proposes a generic architecture for ML systems. He starts by critically analysing the most used architecture for industrial systems, the three-layered architecture. The problems are outlined and his proposal is justified on top of them. He also supports his work with a clear and practical example. I have used his architecture to build a ML system and it was very useful.